+++ eslint.config.js
@@ -1,0 +1,46 @@
+import globals from 'globals';
+import js from '@eslint/js';
+import ts from 'typescript-eslint';
+import vue from 'eslint-plugin-vue';
+import prettier from 'eslint-plugin-prettier/recommended';
+
+export default [
+  {
+    languageOptions: {
+      globals: {
+        ...globals.browser,
+        ...globals.node,
+      },
+    },
+  },
+
+  // js
+  js.configs.recommended,
+
+  // ts
+  ...ts.configs.recommended,
+  {
+    rules: {
+      '@typescript-eslint/no-explicit-any': 'warn',
+    },
+  },
+
+  // vue
+  ...vue.configs['flat/recommended'],
+  {
+    files: ['*.vue', '**/*.vue'],
+    languageOptions: {
+      parserOptions: {
+        parser: ts.parser,
+      },
+    },
+  },
+  {
+    rules: {
+      'vue/multi-word-component-names': 'off',
+    },
+  },
+
+  // prettier
+  prettier,
+];