+++ src/components/bases/FormInput.vue
@@ -1,10 +1,10 @@
 <script setup lang="ts">
-const model = defineModel()
+const model = defineModel() as unknown as string | number | undefined
 </script>
 
 <template>
   <input
-    class="flex-grow border border-black bg-white disabled:cursor-not-allowed disabled:bg-neutral-100 dark:border-white dark:bg-black dark:disabled:bg-neutral-700"
     v-model="model"
+    class="flex-grow border border-black bg-white disabled:cursor-not-allowed disabled:bg-neutral-100 dark:border-white dark:bg-black dark:disabled:bg-neutral-700"
   />
 </template>